* Add OP_LOADP & OP_STOREP to CCustomExplosionGenerator to pass around   pointers in a platform independent way. * Fix indentation of 4 lines in ExplosionGenerator.cpp;   can we please use exactly 1 tab per indentation level? * Convert configHandler.SetInt and .GetInt to signed int   (why was it unsigned in the first place?) * Fix various gcc warnings.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1704 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Projectiles/ExplosiveProjectile.cpp
@@ -79,10 +79,10 @@ void CExplosiveProjectile::Draw(void)
 	dir.Normalize();
 
 	for(int a=0;a<5;++a){
-		col[0]=(5-a)*0.2*weaponDef->visuals.color.x*255;
-		col[1]=(5-a)*0.2*weaponDef->visuals.color.y*255;
-		col[2]=(5-a)*0.2*weaponDef->visuals.color.z*255;
-		col[3]=(5-a)*0.2*weaponDef->intensity*255;
+		col[0]=int((5-a)*0.2*weaponDef->visuals.color.x*255);
+		col[1]=int((5-a)*0.2*weaponDef->visuals.color.y*255);
+		col[2]=int((5-a)*0.2*weaponDef->visuals.color.z*255);
+		col[3]=int((5-a)*0.2*weaponDef->intensity*255);
 		float3 interPos=pos+speed*gu->timeOffset-dir*drawRadius*0.6*a;
 		va->AddVertexTC(interPos-camera->right*drawRadius-camera->up*drawRadius,weaponDef->visuals.texture1->xstart,weaponDef->visuals.texture1->ystart,col);
 		va->AddVertexTC(interPos+camera->right*drawRadius-camera->up*drawRadius,weaponDef->visuals.texture1->xend,weaponDef->visuals.texture1->ystart,col);
